misc/HtmlTestSuite/results/test-0664.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><link href="https://drafts.csswg.org/css-flexbox/#propdef-align-content" rel="help"/>
<title>css-flexbox: Tests align-content with flex-wrap: wrap</title>
<style>
.flexbox {
    display: flex;
    background-color: #aaa;
    position: relative;
    flex-wrap: wrap;
}

.horizontal {
    width: 200px;
    height: 120px;
}
.horizontal &gt; div {
    min-height: 20px;
}
.horizontal &gt; :nth-child(1) {
    width: 100px;
    min-height: 10px;
}
.horizontal &gt; :nth-child(2) {
    width: 100px;
}
.horizontal &gt; :nth-child(3) {
    width: 200px;
}
.horizontal &gt; :nth-child(4) {
    width: 50px;
}

.vertical-rl {
    writing-mode: vertical-rl;
    width: 120px;
    height: 20px;
}
.vertical-rl &gt; div {
    min-width: 20px;
}
.vertical-rl &gt; :nth-child(1) {
    height: 10px;
    min-width: 10px;
}
.vertical-rl &gt; :nth-child(2) {
    height: 10px;
}
.vertical-rl &gt; :nth-child(3) {
    height: 20px;
}
.vertical-rl &gt; :nth-child(4) {
    height: 5px;
}

.flexbox :nth-child(1) {
    background-color: lightblue;
}
.flexbox :nth-child(2) {
    background-color: lightgreen;
}
.flexbox :nth-child(3) {
    background-color: pink;
}
.flexbox :nth-child(4) {
    background-color: yellow;
}
</style>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="/resources/check-layout-th.js"/>
</head><body onload="checkLayout('.flexbox')">
<div id="log"/>
<p>Test to make sure that align-content works properly.</p>

<div class="flexbox horizontal" data-expected-height="120">
  <div data-expected-height="40" data-offset-x="0" data-offset-y="0"/>
  <div data-expected-height="40" data-offset-x="100" data-offset-y="0"/>
  <div data-expected-height="40" data-offset-x="0" data-offset-y="40"/>
  <div data-expected-height="40" data-offset-x="0" data-offset-y="80"/>
</div>

<div class="flexbox horizontal" data-expected-height="120" style="align-content: flex-start">
  <div data-expected-height="20" data-offset-x="0" data-offset-y="0"/>
  <div data-expected-height="20" data-offset-x="100" data-offset-y="0"/>
  <div data-expected-height="20" data-offset-x="0" data-offset-y="20"/>
  <div data-expected-height="20" data-offset-x="0" data-offset-y="40"/>
</div>

<div class="flexbox horizontal" data-expected-height="120" style="align-content: flex-end">
  <div data-expected-height="20" data-offset-x="0" data-offset-y="60"/>
  <div data-expected-height="20" data-offset-x="100" data-offset-y="60"/>
  <div data-expected-height="20" data-offset-x="0" data-offset-y="80"/>
  <div data-expected-height="20" data-offset-x="0" data-offset-y="100"/>
</div>

<div class="flexbox horizontal" data-expected-height="120" style="align-content: center">
  <div data-expected-height="20" data-offset-x="0" data-offset-y="30"/>
  <div data-expected-height="20" data-offset-x="100" data-offset-y="30"/>
  <div data-expected-height="20" data-offset-x="0" data-offset-y="50"/>
  <div data-expected-height="20" data-offset-x="0" data-offset-y="70"/>
</div>

<div class="flexbox horizontal" data-expected-height="120" style="align-content: space-between">
  <div data-expected-height="20" data-offset-x="0" data-offset-y="0"/>
  <div data-expected-height="20" data-offset-x="100" data-offset-y="0"/>
  <div data-expected-height="20" data-offset-x="0" data-offset-y="50"/>
  <div data-expected-height="20" data-offset-x="0" data-offset-y="100"/>
</div>

<div class="flexbox horizontal" data-expected-height="120" style="align-content: space-evenly">
  <div data-expected-height="20" data-offset-x="0" data-offset-y="15"/>
  <div data-expected-height="20" data-offset-x="100" data-offset-y="15"/>
  <div data-expected-height="20" data-offset-x="0" data-offset-y="50"/>
  <div data-expected-height="20" data-offset-x="0" data-offset-y="85"/>
</div>

<div class="flexbox horizontal" data-expected-height="120" style="align-content: space-around">
  <div data-expected-height="20" data-offset-x="0" data-offset-y="10"/>
  <div data-expected-height="20" data-offset-x="100" data-offset-y="10"/>
  <div data-expected-height="20" data-offset-x="0" data-offset-y="50"/>
  <div data-expected-height="20" data-offset-x="0" data-offset-y="90"/>
</div>

<div class="flexbox horizontal" data-expected-height="120" style="align-content: stretch">
  <div data-expected-height="40" data-offset-x="0" data-offset-y="0"/>
  <div data-expected-height="40" data-offset-x="100" data-offset-y="0"/>
  <div data-expected-height="40" data-offset-x="0" data-offset-y="40"/>
  <div data-expected-height="40" data-offset-x="0" data-offset-y="80"/>
</div>

<!-- Negative overflow goes out the top. -->
<div class="flexbox horizontal" data-expected-height="30" style="align-content: flex-end; height: 30px">
  <div data-expected-height="20" data-offset-x="0" data-offset-y="-30"/>
  <div data-expected-height="20" data-offset-x="100" data-offset-y="-30"/>
  <div data-expected-height="20" data-offset-x="0" data-offset-y="-10"/>
  <div data-expected-height="20" data-offset-x="0" data-offset-y="10"/>
</div>

<!-- If we overflow, we should true center. -->
<div class="flexbox horizontal" data-expected-height="30" style="align-content: center; height: 30px">
  <div data-expected-height="20" data-offset-x="0" data-offset-y="-15"/>
  <div data-expected-height="20" data-offset-x="100" data-offset-y="-15"/>
  <div data-expected-height="20" data-offset-x="0" data-offset-y="5"/>
  <div data-expected-height="20" data-offset-x="0" data-offset-y="25"/>
</div>

<!-- If we overflow, we should be the same as flex-start. -->
<div class="flexbox horizontal" data-expected-height="30" style="align-content: space-between; height: 30px">
  <div data-expected-height="20" data-offset-x="0" data-offset-y="0"/>
  <div data-expected-height="20" data-offset-x="100" data-offset-y="0"/>
  <div data-expected-height="20" data-offset-x="0" data-offset-y="20"/>
  <div data-expected-height="20" data-offset-x="0" data-offset-y="40"/>
</div>

<!-- If we overflow, we should true center. -->
<div class="flexbox horizontal" data-expected-height="30" style="align-content: space-around; height: 30px">
  <div data-expected-height="20" data-offset-x="0" data-offset-y="-15"/>
  <div data-expected-height="20" data-offset-x="100" data-offset-y="-15"/>
  <div data-expected-height="20" data-offset-x="0" data-offset-y="5"/>
  <div data-expected-height="20" data-offset-x="0" data-offset-y="25"/>
</div>

<!-- If we overflow, we should true center. -->
<div class="flexbox horizontal" data-expected-height="30" style="align-content: space-evenly; height: 30px">
  <div data-expected-height="20" data-offset-x="0" data-offset-y="-15"/>
  <div data-expected-height="20" data-offset-x="100" data-offset-y="-15"/>
  <div data-expected-height="20" data-offset-x="0" data-offset-y="5"/>
  <div data-expected-height="20" data-offset-x="0" data-offset-y="25"/>
</div>

<!-- Stretch should only grow, not shrink. -->
<div class="flexbox horizontal" data-expected-height="30" style="align-content: stretch; height: 30px">
  <div data-expected-height="20" data-offset-x="0" data-offset-y="0"/>
  <div data-expected-height="20" data-offset-x="100" data-offset-y="0"/>
  <div data-expected-height="20" data-offset-x="0" data-offset-y="20"/>
  <div data-expected-height="20" data-offset-x="0" data-offset-y="40"/>
</div>

<!-- 0 lines should not crash. -->
<div class="flexbox horizontal" data-expected-height="30" style="align-content: space-between; height: 30px">
</div>
<div class="flexbox horizontal" data-expected-height="30" style="align-content: space-around; height: 30px">
</div>
<div class="flexbox horizontal" data-expected-height="30" style="align-content: space-evenly; height: 30px">
</div>
<div class="flexbox horizontal" data-expected-height="30" style="align-content: stretch; height: 30px">
</div>

<!-- 1 line should not crash. -->
<div class="flexbox horizontal" data-expected-height="30" style="align-content: space-between; height: 30px">
  <div data-expected-height="10" data-offset-x="0" data-offset-y="0"/>
</div>
<div class="flexbox horizontal" data-expected-height="30" style="align-content: space-around; height: 30px">
  <div data-expected-height="10" data-offset-x="0" data-offset-y="10"/>
</div>
<div class="flexbox horizontal" data-expected-height="30" style="align-content: space-evenly; height: 30px">
  <div data-expected-height="10" data-offset-x="0" data-offset-y="10"/>
</div>
<div class="flexbox horizontal" data-expected-height="30" style="align-content: stretch; height: 30px">
  <div data-expected-height="30" data-offset-x="0" data-offset-y="0"/>
</div>


<div class="flexbox vertical-rl" data-expected-width="120">
  <div data-expected-width="40" data-offset-x="80" data-offset-y="0"/>
  <div data-expected-width="40" data-offset-x="80" data-offset-y="10"/>
  <div data-expected-width="40" data-offset-x="40" data-offset-y="0"/>
  <div data-expected-width="40" data-offset-x="0" data-offset-y="0"/>
</div>

<div class="flexbox vertical-rl" data-expected-width="120" style="align-content: flex-start">
  <div data-expected-width="20" data-offset-x="100" data-offset-y="0"/>
  <div data-expected-width="20" data-offset-x="100" data-offset-y="10"/>
  <div data-expected-width="20" data-offset-x="80" data-offset-y="0"/>
  <div data-expected-width="20" data-offset-x="60" data-offset-y="0"/>
</div>

<div class="flexbox vertical-rl" data-expected-width="120" style="align-content: flex-end">
  <div data-expected-width="20" data-offset-x="40" data-offset-y="0"/>
  <div data-expected-width="20" data-offset-x="40" data-offset-y="10"/>
  <div data-expected-width="20" data-offset-x="20" data-offset-y="0"/>
  <div data-expected-width="20" data-offset-x="0" data-offset-y="0"/>
</div>

<div class="flexbox vertical-rl" data-expected-width="120" style="align-content: center">
  <div data-expected-width="20" data-offset-x="70" data-offset-y="0"/>
  <div data-expected-width="20" data-offset-x="70" data-offset-y="10"/>
  <div data-expected-width="20" data-offset-x="50" data-offset-y="0"/>
  <div data-expected-width="20" data-offset-x="30" data-offset-y="0"/>
</div>

<div class="flexbox vertical-rl" data-expected-width="120" style="align-content: space-between">
  <div data-expected-width="20" data-offset-x="100" data-offset-y="0"/>
  <div data-expected-width="20" data-offset-x="100" data-offset-y="10"/>
  <div data-expected-width="20" data-offset-x="50" data-offset-y="0"/>
  <div data-expected-width="20" data-offset-x="0" data-offset-y="0"/>
</div>

<div class="flexbox vertical-rl" data-expected-width="120" style="align-content: space-around">
  <div data-expected-width="20" data-offset-x="90" data-offset-y="0"/>
  <div data-expected-width="20" data-offset-x="90" data-offset-y="10"/>
  <div data-expected-width="20" data-offset-x="50" data-offset-y="0"/>
  <div data-expected-width="20" data-offset-x="10" data-offset-y="0"/>
</div>

<div class="flexbox vertical-rl" data-expected-width="120" style="align-content: space-evenly">
  <div data-expected-width="20" data-offset-x="85" data-offset-y="0"/>
  <div data-expected-width="20" data-offset-x="85" data-offset-y="10"/>
  <div data-expected-width="20" data-offset-x="50" data-offset-y="0"/>
  <div data-expected-width="20" data-offset-x="15" data-offset-y="0"/>
</div>

<div class="flexbox vertical-rl" data-expected-width="120" style="align-content: stretch">
  <div data-expected-width="40" data-offset-x="80" data-offset-y="0"/>
  <div data-expected-width="40" data-offset-x="80" data-offset-y="10"/>
  <div data-expected-width="40" data-offset-x="40" data-offset-y="0"/>
  <div data-expected-width="40" data-offset-x="0" data-offset-y="0"/>
</div>

<!-- Negative overflow goes out the right. -->
<div class="flexbox vertical-rl" data-expected-width="30" style="align-content: flex-end; width: 30px;">
  <div data-expected-width="20" data-offset-x="40" data-offset-y="0"/>
  <div data-expected-width="20" data-offset-x="40" data-offset-y="10"/>
  <div data-expected-width="20" data-offset-x="20" data-offset-y="0"/>
  <div data-expected-width="20" data-offset-x="0" data-offset-y="0"/>
</div>

<!-- If we overflow, we should true center. -->
<div class="flexbox vertical-rl" data-expected-width="30" style="align-content: center; width: 30px;">
  <div data-expected-width="20" data-offset-x="25" data-offset-y="0"/>
  <div data-expected-width="20" data-offset-x="25" data-offset-y="10"/>
  <div data-expected-width="20" data-offset-x="5" data-offset-y="0"/>
  <div data-expected-width="20" data-offset-x="-15" data-offset-y="0"/>
</div>

<!-- If we overflow, we should be the same as flex-start. -->
<div class="flexbox vertical-rl" data-expected-width="30" style="align-content: space-between; width: 30px;">
  <div data-expected-width="20" data-offset-x="10" data-offset-y="0"/>
  <div data-expected-width="20" data-offset-x="10" data-offset-y="10"/>
  <div data-expected-width="20" data-offset-x="-10" data-offset-y="0"/>
  <div data-expected-width="20" data-offset-x="-30" data-offset-y="0"/>
</div>

<!-- If we overflow, we should true center. -->
<div class="flexbox vertical-rl" data-expected-width="30" style="align-content: space-around; width: 30px;">
  <div data-expected-width="20" data-offset-x="25" data-offset-y="0"/>
  <div data-expected-width="20" data-offset-x="25" data-offset-y="10"/>
  <div data-expected-width="20" data-offset-x="5" data-offset-y="0"/>
  <div data-expected-width="20" data-offset-x="-15" data-offset-y="0"/>
</div>

<!-- If we overflow, we should true center. -->
<div class="flexbox vertical-rl" data-expected-width="30" style="align-content: space-evenly; width: 30px;">
  <div data-expected-width="20" data-offset-x="25" data-offset-y="0"/>
  <div data-expected-width="20" data-offset-x="25" data-offset-y="10"/>
  <div data-expected-width="20" data-offset-x="5" data-offset-y="0"/>
  <div data-expected-width="20" data-offset-x="-15" data-offset-y="0"/>
</div>

<!-- Stretch should only grow, not shrink. -->
<div class="flexbox vertical-rl" data-expected-width="30" style="align-content: stretch; width: 30px;">
  <div data-expected-width="20" data-offset-x="10" data-offset-y="0"/>
  <div data-expected-width="20" data-offset-x="10" data-offset-y="10"/>
  <div data-expected-width="20" data-offset-x="-10" data-offset-y="0"/>
  <div data-expected-width="20" data-offset-x="-30" data-offset-y="0"/>
</div>

<!-- 0 lines should not crash. -->
<div class="flexbox vertical-rl" data-expected-width="30" style="align-content: space-between; width: 30px">
</div>
<div class="flexbox vertical-rl" data-expected-width="30" style="align-content: space-around; width: 30px">
</div>
<div class="flexbox vertical-rl" data-expected-width="30" style="align-content: space-evenly; width: 30px">
</div>
<div class="flexbox vertical-rl" data-expected-width="30" style="align-content: stretch; width: 30px">
</div>

<!-- 1 line should not crash. -->
<div class="flexbox vertical-rl" data-expected-width="30" style="align-content: space-between; width: 30px">
  <div data-expected-width="10" data-offset-x="20" data-offset-y="0"/>
</div>
<div class="flexbox vertical-rl" data-expected-width="30" style="align-content: space-around; width: 30px">
  <div data-expected-width="10" data-offset-x="10" data-offset-y="0"/>
</div>
<div class="flexbox vertical-rl" data-expected-width="30" style="align-content: space-evenly; width: 30px">
  <div data-expected-width="10" data-offset-x="10" data-offset-y="0"/>
</div>
<div class="flexbox vertical-rl" data-expected-width="30" style="align-content: stretch; width: 30px">
  <div data-expected-width="30" data-offset-x="0" data-offset-y="0"/>
</div>
</body></html>
